--- modules/parametric.py
@@ -1,6 +1,0 @@
-from modules.truth_info import TruthSource
-filepath_truth = "/scratch/etolley/SDC2/dev/sky_ldev_truthcat_v2.txt"
-filpath_trainingdata = "/scratch/etolley/SDC2/dev_training_denoised/"
-
-
-
