+++ pipelines/run_SDC2_ldev_pipeline.sbatch
@@ -11,7 +11,7 @@ echo STARTING AT `date`
 module load intel/18.0.2
 module load python/3.6.5
 module load cmake/3.11.1 # for pysap
-source venvs/sdc2/bin/activate
+source venvs/lisa/bin/activate
 export PYTHONPATH="$PWD" 
 
 srun python pipelines/pipeline.py pipelines/SDC2_ldev.config 144 $SLURM_ARRAY_TASK_ID
